Clean up str/bytes/unicode encoding in C.
Minor perf improvements included.

+++ bson/_cbsonmodule.h
@@ -37,7 +37,7 @@ typedef int Py_ssize_t;
 
 #define _cbson_write_pair_INDEX 2
 #define _cbson_write_pair_RETURN int
-#define _cbson_write_pair_PROTO (PyObject* self, buffer_t buffer, const char* name, Py_ssize_t name_length, PyObject* value, unsigned char check_keys, unsigned char uuid_subtype, unsigned char allow_id)
+#define _cbson_write_pair_PROTO (PyObject* self, buffer_t buffer, const char* name, int name_length, PyObject* value, unsigned char check_keys, unsigned char uuid_subtype, unsigned char allow_id)
 
 #define _cbson_decode_and_write_pair_INDEX 3
 #define _cbson_decode_and_write_pair_RETURN int
